GPX File from back to front
GPX File stored as BYTEA data in postgre dB. Pulled through as base64String binary from the backend via controller.

The log entry class now contains methods to decode this gpxFile and get a URL for it that can be plotted in the LeafletGPXMap. This needs refactoring for responsibility purposes but works well.